feat(publish): add hidden runt publish command
