+++ crates/runt/Cargo.toml
@@ -15,6 +15,7 @@ anyhow = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 uuid = { workspace = true }
+runt-publish = { path = "../runt-publish" }
 runtimed-client = { path = "../runtimed-client" }
 runtimed-service = { path = "../runtimed-service" }
 nteract-telemetry = { path = "../nteract-telemetry" }
